revert ota_base changes, move to platform
