+++ esphome/components/micro_wake_word/micro_wake_word.cpp
@@ -9,7 +9,7 @@
 #include "esphome/components/audio/audio_transfer_buffer.h"
 
 #ifdef USE_OTA
-#include "esphome/components/ota_base/ota_backend.h"
+#include "esphome/components/ota/ota_backend.h"
 #endif
 
 namespace esphome {
@@ -121,11 +121,11 @@ void MicroWakeWord::setup() {
   });
 
 #ifdef USE_OTA
-  ota_base::get_global_ota_callback()->add_on_state_callback(
-      [this](ota_base::OTAState state, float progress, uint8_t error, ota_base::OTAComponent *comp) {
-        if (state == ota_base::OTA_STARTED) {
+  ota::get_global_ota_callback()->add_on_state_callback(
+      [this](ota::OTAState state, float progress, uint8_t error, ota::OTAComponent *comp) {
+        if (state == ota::OTA_STARTED) {
           this->suspend_task_();
-        } else if (state == ota_base::OTA_ERROR) {
+        } else if (state == ota::OTA_ERROR) {
           this->resume_task_();
         }
       });
