+++ esphome/components/ota/automation.h
@@ -1,15 +1,11 @@
 #pragma once
 #ifdef USE_OTA_STATE_CALLBACK
-#include "esphome/components/ota_base/ota_backend.h"
+#include "esphome/components/ota/ota_backend.h"
 
 #include "esphome/core/automation.h"
 
 namespace esphome {
 namespace ota {
-
-// Import types from ota_base for the automation triggers
-using ota_base::OTAComponent;
-using ota_base::OTAState;
 
 class OTAStateChangeTrigger : public Trigger<OTAState> {
  public:
@@ -26,7 +22,7 @@ class OTAStartTrigger : public Trigger<> {
  public:
   explicit OTAStartTrigger(OTAComponent *parent) {
     parent->add_on_state_callback([this, parent](OTAState state, float progress, uint8_t error) {
-      if (state == ota_base::OTA_STARTED && !parent->is_failed()) {
+      if (state == ota::OTA_STARTED && !parent->is_failed()) {
         trigger();
       }
     });
@@ -37,7 +33,7 @@ class OTAProgressTrigger : public Trigger<float> {
  public:
   explicit OTAProgressTrigger(OTAComponent *parent) {
     parent->add_on_state_callback([this, parent](OTAState state, float progress, uint8_t error) {
-      if (state == ota_base::OTA_IN_PROGRESS && !parent->is_failed()) {
+      if (state == ota::OTA_IN_PROGRESS && !parent->is_failed()) {
         trigger(progress);
       }
     });
@@ -48,7 +44,7 @@ class OTAEndTrigger : public Trigger<> {
  public:
   explicit OTAEndTrigger(OTAComponent *parent) {
     parent->add_on_state_callback([this, parent](OTAState state, float progress, uint8_t error) {
-      if (state == ota_base::OTA_COMPLETED && !parent->is_failed()) {
+      if (state == ota::OTA_COMPLETED && !parent->is_failed()) {
         trigger();
       }
     });
@@ -59,7 +55,7 @@ class OTAAbortTrigger : public Trigger<> {
  public:
   explicit OTAAbortTrigger(OTAComponent *parent) {
     parent->add_on_state_callback([this, parent](OTAState state, float progress, uint8_t error) {
-      if (state == ota_base::OTA_ABORT && !parent->is_failed()) {
+      if (state == ota::OTA_ABORT && !parent->is_failed()) {
         trigger();
       }
     });
@@ -70,7 +66,7 @@ class OTAErrorTrigger : public Trigger<uint8_t> {
  public:
   explicit OTAErrorTrigger(OTAComponent *parent) {
     parent->add_on_state_callback([this, parent](OTAState state, float progress, uint8_t error) {
-      if (state == ota_base::OTA_ERROR && !parent->is_failed()) {
+      if (state == ota::OTA_ERROR && !parent->is_failed()) {
         trigger(error);
       }
     });
