+++ esphome/components/ota/ota_backend_arduino_esp32.cpp
@@ -1,20 +1,20 @@
-#ifdef USE_LIBRETINY
-#include "ota_backend_arduino_libretiny.h"
-#include "ota_backend.h"
-
+#ifdef USE_ESP32_FRAMEWORK_ARDUINO
 #include "esphome/core/defines.h"
 #include "esphome/core/log.h"
+
+#include "ota_backend.h"
+#include "ota_backend_arduino_esp32.h"
 
 #include <Update.h>
 
 namespace esphome {
-namespace ota_base {
+namespace ota {
 
-static const char *const TAG = "ota.arduino_libretiny";
+static const char *const TAG = "ota.arduino_esp32";
 
-std::unique_ptr<OTABackend> make_ota_backend() { return make_unique<ArduinoLibreTinyOTABackend>(); }
+std::unique_ptr<OTABackend> make_ota_backend() { return make_unique<ArduinoESP32OTABackend>(); }
 
-OTAResponseTypes ArduinoLibreTinyOTABackend::begin(size_t image_size) {
+OTAResponseTypes ArduinoESP32OTABackend::begin(size_t image_size) {
   // Handle UPDATE_SIZE_UNKNOWN (0) which is used by web server OTA
   // where the exact firmware size is unknown due to multipart encoding
   if (image_size == 0) {
@@ -34,12 +34,12 @@ OTAResponseTypes ArduinoLibreTinyOTABackend::begin(size_t image_size) {
   return OTA_RESPONSE_ERROR_UNKNOWN;
 }
 
-void ArduinoLibreTinyOTABackend::set_update_md5(const char *md5) {
+void ArduinoESP32OTABackend::set_update_md5(const char *md5) {
   Update.setMD5(md5);
   this->md5_set_ = true;
 }
 
-OTAResponseTypes ArduinoLibreTinyOTABackend::write(uint8_t *data, size_t len) {
+OTAResponseTypes ArduinoESP32OTABackend::write(uint8_t *data, size_t len) {
   size_t written = Update.write(data, len);
   if (written == len) {
     return OTA_RESPONSE_OK;
@@ -51,7 +51,7 @@ OTAResponseTypes ArduinoLibreTinyOTABackend::write(uint8_t *data, size_t len) {
   return OTA_RESPONSE_ERROR_WRITING_FLASH;
 }
 
-OTAResponseTypes ArduinoLibreTinyOTABackend::end() {
+OTAResponseTypes ArduinoESP32OTABackend::end() {
   // Use strict validation (false) when MD5 is set, lenient validation (true) when no MD5
   // This matches the behavior of the old web_server OTA implementation
   if (Update.end(!this->md5_set_)) {
@@ -64,9 +64,9 @@ OTAResponseTypes ArduinoLibreTinyOTABackend::end() {
   return OTA_RESPONSE_ERROR_UPDATE_END;
 }
 
-void ArduinoLibreTinyOTABackend::abort() { Update.abort(); }
+void ArduinoESP32OTABackend::abort() { Update.abort(); }
 
-}  // namespace ota_base
+}  // namespace ota
 }  // namespace esphome
 
-#endif  // USE_LIBRETINY
+#endif  // USE_ESP32_FRAMEWORK_ARDUINO
