+++ esphome/components/ota/ota_backend_arduino_libretiny.h
@@ -1,16 +1,13 @@
 #pragma once
-#ifdef USE_ESP_IDF
+#ifdef USE_LIBRETINY
 #include "ota_backend.h"
 
-#include "esphome/components/md5/md5.h"
 #include "esphome/core/defines.h"
 
-#include <esp_ota_ops.h>
-
 namespace esphome {
-namespace ota_base {
+namespace ota {
 
-class IDFOTABackend : public OTABackend {
+class ArduinoLibreTinyOTABackend : public OTABackend {
  public:
   OTAResponseTypes begin(size_t image_size) override;
   void set_update_md5(const char *md5) override;
@@ -20,13 +17,10 @@ class IDFOTABackend : public OTABackend {
   bool supports_compression() override { return false; }
 
  private:
-  esp_ota_handle_t update_handle_{0};
-  const esp_partition_t *partition_;
-  md5::MD5Digest md5_{};
-  char expected_bin_md5_[32];
   bool md5_set_{false};
 };
 
-}  // namespace ota_base
+}  // namespace ota
 }  // namespace esphome
-#endif
+
+#endif  // USE_LIBRETINY
