+++ esphome/components/ota/ota_backend_arduino_rp2040.h
@@ -1,14 +1,15 @@
 #pragma once
-#ifdef USE_ESP32_FRAMEWORK_ARDUINO
+#ifdef USE_ARDUINO
+#ifdef USE_RP2040
 #include "ota_backend.h"
 
 #include "esphome/core/defines.h"
-#include "esphome/core/helpers.h"
+#include "esphome/core/macros.h"
 
 namespace esphome {
-namespace ota_base {
+namespace ota {
 
-class ArduinoESP32OTABackend : public OTABackend {
+class ArduinoRP2040OTABackend : public OTABackend {
  public:
   OTAResponseTypes begin(size_t image_size) override;
   void set_update_md5(const char *md5) override;
@@ -21,7 +22,8 @@ class ArduinoESP32OTABackend : public OTABackend {
   bool md5_set_{false};
 };
 
-}  // namespace ota_base
+}  // namespace ota
 }  // namespace esphome
 
-#endif  // USE_ESP32_FRAMEWORK_ARDUINO
+#endif  // USE_RP2040
+#endif  // USE_ARDUINO
